Fix bug in serializing researchers_paper object
Previously the program will not store information of which paper
was downloaded last.

So when main_download_paper() is run the next day,
program will redownload a few papers of the same author that were
already downloaded.
This bug fixes that, and will only serialize the papers that
indeed have not been downloaded